@@ -4,9 +4,14 @@
   "description": "Cloud Pricing API",
   "main": "src/server.js",
   "scripts": {
-    "update": "node src/update.js",
-    "start": "node src/server.js",
-    "lint": "eslint --ext .js --ignore-path .gitignore ."
+    "dev": "nodemon src/server.ts",
+    "build": "tsc",
+    "start": "node dist/server.js",
+    "update": "per-env",
+    "update:development": "ts-node src/cmd/update.ts",
+    "update:production": "node dist/cmd/update.js",
+    "lint": "eslint --ext .ts,.js --ignore-path .gitignore .",
+    "format": "prettier --write 'src/**/*.{js,ts}'"
   },
   "author": "Alistair Scott",
   "license": "Apache-2.0",
@@ -18,26 +23,43 @@
   ],
   "dependencies": {
     "apollo-server-express": "^2.17.0",
+    "apollo-server-plugin-base": "^0.10.2",
     "axios": "^0.19.2",
     "dotenv": "^8.2.0",
     "express": "^4.17.1",
     "glob": "^7.1.6",
+    "graphql": "^15.0.0",
+    "graphql-parse-resolve-info": "^4.9.0",
+    "graphql-tools": "^6.2.4",
     "lodash": "^4.17.19",
     "mingo": "^3.0.5",
     "mongodb": "^3.5.8",
+    "per-env": "^1.0.2",
     "pino": "^6.3.2",
     "pino-http": "^5.2.0",
     "pino-pretty": "^4.0.0",
     "prettier": "^2.1.1",
-    "yargs": "^15.4.1",
-    "graphql": "^14.7.0"
+    "yargs": "^15.4.1"
   },
   "devDependencies": {
     "@babel/core": "^7.10.5",
     "@babel/node": "^7.10.5",
-    "eslint": "^6.8.0",
+    "@types/glob": "^7.1.3",
+    "@types/lodash": "^4.14.161",
+    "@types/mongodb": "^3.5.33",
+    "@types/node": "^14.14.6",
+    "@types/pino-http": "^5.0.5",
+    "@types/prettier": "^2.1.1",
+    "@types/yargs": "^15.0.7",
+    "@typescript-eslint/eslint-plugin": "^4.3.0",
+    "@typescript-eslint/parser": "^4.3.0",
+    "eslint": "^7.2.0",
     "eslint-config-airbnb-base": "^14.1.0",
-    "eslint-plugin-import": "^2.21.1"
+    "eslint-config-prettier": "^6.12.0",
+    "eslint-plugin-import": "^2.21.1",
+    "nodemon": "^2.0.6",
+    "ts-node": "^9.0.0",
+    "typescript": "^4.0.3"
   },
   "repository": {
     "type": "git",
