@@ -1,8 +1,31 @@
 module.exports = {
-  extends: 'airbnb-base',
+  env: {
+    node: true,
+  },
+  extends: [
+    'airbnb-base',
+    'plugin:@typescript-eslint/recommended',
+    'prettier',
+  ],
+  parser: '@typescript-eslint/parser',
+  plugins: ['@typescript-eslint'],
+  settings: {
+    'import/resolver': {
+      node: {
+        paths: ['src'],
+        extensions: ['.ts', '.js'],
+      },
+    },
+  },
   rules: {
+    'import/extensions': 'off',
+    'max-classes-per-file': 'off',
     'max-len': 'off',
+    'no-await-in-loop': 'off',
     'no-continue': 'off',
+    'no-plusplus': 'off',
     'no-restricted-syntax': 'off',
+    'no-use-before-define': 'off',
+    'no-useless-constructor': 'off',
   },
 };
